requirements/dev.txt

asgiref==3.2.10
cffi==1.14.3
cryptography==2.9.2
Django==3.1.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
jwt==1.0.0
psycopg2==2.8.6
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
